Fix: C++11 does not allow narrowing in brace list initializers.

+++ src/ManagerWindows.cc
@@ -401,7 +401,10 @@ RootWO::setEwmhDesktopLayout(void)
     // This property is defined to be set by the pager, however, as pekwm
     // displays a "pager" when changing workspaces and other applications
     // might want to read this information set it anyway.
-    long desktop_layout[] = { NET_WM_ORIENTATION_HORZ, Workspaces::getPerRow(), Workspaces::getRows(), NET_WM_TOPLEFT };
+    long desktop_layout[] = { NET_WM_ORIENTATION_HORZ,
+                              static_cast<long>(Workspaces::getPerRow()),
+                              static_cast<long>(Workspaces::getRows()),
+                              NET_WM_TOPLEFT };
     X11::setLongs(X11::getRoot(), NET_DESKTOP_LAYOUT, desktop_layout, sizeof(desktop_layout)/sizeof(desktop_layout[0]));
 }
